Add files via upload. Update biography and main.
Added files socialmedia.html and socialmedia.css to present a page with links to my social media accounts. It is presented by showing easy to recognize icons. I also added resume.html and resume.css which display a page that contains a pdf file viewer that displays my resume. The main and biography files were simply updated to change the ordering of the links in the header.